Replaced parseInt() with parseFloat() to correctly handle decimals Replaced all instances of parseInt() to parseFloat() to correctly handle decimals. Didn't realize the former truncates decimals as well which is why decimals weren't being handled properly.

+++ script.js
@@ -78,20 +78,20 @@ operateButton.addEventListener("click", () =>
         display.textContent = result);
 
 function add (a, b) {
-  return parseInt(a) + parseInt(b);
+  return parseFloat(a) + parseFloat(b);
 
 };
 
 function subtract (a, b) {
-    return parseInt(a) - parseInt(b);
+    return parseFloat(a) - parseFloat(b);
 };
 
 function multiply (a, b) {
-    return parseInt(a) * parseInt(b);
+    return parseFloat(a) * parseFloat(b);
 };
 
 function divide (a, b) {
-     return parseInt(a) / parseInt(b);
+     return parseFloat(a) / parseFloat(b);
 };
 
 function operate () {
